added a few functionalities, including fetching a specific holding with the user_id and the ticker.

+++ src/models/holding.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, Numeric
+from sqlalchemy import Column, ForeignKey, Integer, String, Numeric, Float
 
 from src.models.base import Base
 
@@ -9,5 +9,5 @@ class Holding(Base):
     id = Column(Integer, primary_key=True)
     user_id = Column(Integer, ForeignKey("users.id"))
     ticker = Column(String)
-    quantity = Column(Integer)
+    quantity = Column(Float)
     buy_in_price = Column(Numeric(10, 2))
